+++ providers/defltprov.c
@@ -650,7 +650,7 @@ static const OSSL_ALGORITHM deflt_keymgmt[] = {
         PROV_DESCS_curveSM2 },
 #endif
 #ifndef OPENSSL_NO_LMS
-    { PROV_NAMES_LMS, "provider=default", ossl_lms_keymgmt_functions,
+    { PROV_NAMES_LMS, "provider=default", ossl_hss_lms_keymgmt_functions,
         PROV_DESCS_LMS },
 #endif
 #ifndef OPENSSL_NO_ML_KEM