+++ providers/implementations/encode_decode/decode_der2key.c
@@ -44,7 +44,7 @@
 #include "internal/nelem.h"
 #include "prov/ml_dsa_codecs.h"
 #include "prov/ml_kem_codecs.h"
-#include "prov/lms_codecs.h"
+#include "prov/hss_lms_codecs.h"
 #include "providers/implementations/encode_decode/decode_der2key.inc"
 
 #ifndef OPENSSL_NO_SLH_DSA
@@ -1003,17 +1003,17 @@ static ossl_inline void *ml_dsa_d2i_PUBKEY(const uint8_t **der, long der_len,
 /* ---------------------------------------------------------------------- */
 
 #ifndef OPENSSL_NO_LMS
-#define lms_evp_type EVP_PKEY_HSS_LMS
-#define lms_free (free_key_fn *)ossl_lms_key_free
-#define lms_check NULL
-#define lms_adjust NULL
+#define hss_lms_evp_type EVP_PKEY_HSS_LMS
+#define hss_lms_free (free_key_fn *)ossl_hss_lms_key_free
+#define hss_lms_check NULL
+#define hss_lms_adjust NULL
 
-static ossl_inline void *lms_d2i_PUBKEY(const uint8_t **der, long der_len,
+static ossl_inline void *hss_lms_d2i_PUBKEY(const uint8_t **der, long der_len,
     struct der2key_ctx_st *ctx)
 {
-    LMS_KEY *key;
+    HSS_LMS_KEY *key;
 
-    key = ossl_lms_d2i_PUBKEY(*der, der_len, ctx->provctx);
+    key = ossl_hss_lms_d2i_PUBKEY(*der, der_len, ctx->provctx);
     if (key != NULL)
         *der += der_len;
     return key;
@@ -1325,5 +1325,5 @@ MAKE_DECODER("ML-DSA-87", ml_dsa_87, ml_dsa_87, SubjectPublicKeyInfo);
 #endif
 
 #ifndef OPENSSL_NO_LMS
-MAKE_DECODER("LMS", lms, lms, SubjectPublicKeyInfo);
+MAKE_DECODER("HSS_LMS", hss_lms, hss_lms, SubjectPublicKeyInfo);
 #endif