+++ test/lms_test.c
@@ -7,6 +7,13 @@
  * https://www.openssl.org/source/license.html
  */
 
+/*
+ * Rather than creating a nearly identical file for HSS,
+ * this file is used to test either LMS or HSS public key and signature
+ * verification. It shares the code since the public key encoding only differs
+ * by an additional 4 bytes (for L).
+ */
+
 #include <openssl/core_names.h>
 #include <openssl/decoder.h>
 #include <openssl/evp.h>
@@ -19,6 +26,7 @@ typedef enum OPTION_choice {
     OPT_ERR = -1,
     OPT_EOF = 0,
     OPT_CONFIG_FILE,
+    OPT_KEY_ALG,
     OPT_TEST_ENUM
 } OPTION_CHOICE;
 
@@ -26,8 +34,12 @@ static OSSL_LIB_CTX *libctx = NULL;
 static char *propq = NULL;
 static OSSL_PROVIDER *nullprov = NULL;
 static OSSL_PROVIDER *libprov = NULL;
+/* Choose LMS by default */
+static char *key_alg = "LMS";
+static LMS_ACVP_TEST_DATA *testdata = lms_testdata;
+static size_t testdata_len = OSSL_NELEM(lms_testdata);
 
-static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
+static EVP_PKEY *pubkey_from_data(const unsigned char *data, size_t datalen)
 {
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
@@ -37,7 +49,7 @@ static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
     params[0] = OSSL_PARAM_construct_octet_string(OSSL_PKEY_PARAM_PUB_KEY,
         (unsigned char *)data, datalen);
     params[1] = OSSL_PARAM_construct_end();
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", propq))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, propq))
         && TEST_int_eq(EVP_PKEY_fromdata_init(ctx), 1)
         && (EVP_PKEY_fromdata(ctx, &key, EVP_PKEY_PUBLIC_KEY, params) == 1);
     if (ret == 0) {
@@ -51,7 +63,7 @@ static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
 static int lms_bad_pub_len_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     size_t publen = 0;
     unsigned char pubdata[128];
@@ -65,7 +77,7 @@ static int lms_bad_pub_len_test(void)
     for (publen = 0; publen <= td->publen + 16; publen += 3) {
         if (publen == td->publen)
             continue;
-        if (!TEST_ptr_null(pkey = lms_pubkey_from_data(pubdata, publen)))
+        if (!TEST_ptr_null(pkey = pubkey_from_data(pubdata, publen)))
             goto end;
     }
     ret = 1;
@@ -78,51 +90,80 @@ end:
     return ret == 1;
 }
 
-static int lms_pubkey_decoder_fail_test(void)
+/* params[] contains a list of invalid public key data */
+static int do_pubkey_decoder_fail(const OSSL_PARAM params[])
 {
     int ret = 0;
     EVP_PKEY *pkey = NULL;
     OSSL_DECODER_CTX *dctx = NULL;
     int selection = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
     const unsigned char *pdata;
     size_t pdatalen;
-    static const unsigned char pub_bad_LMSType[] = {
-        0x00, 0x00, 0x00, 0xAA
-    };
-    static const unsigned char pub_bad_OTSType[] = {
-        0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
-    };
+    const OSSL_PARAM *p = params;
 
-    if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pkey, NULL, NULL, "LMS",
-                      selection,
-                      libctx, NULL)))
+    if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pkey, NULL, NULL,
+                      "LMS", selection, libctx, NULL)))
         return 0;
 
-    pdata = td->pub;
-    pdatalen = 3;
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
+    while (p->key != NULL) {
+        pdata = p->data;
+        pdatalen = p->data_size;
 
-    pdatalen = SIZE_MAX;
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
-    pdata = pub_bad_LMSType;
-    pdatalen = sizeof(pub_bad_LMSType);
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
-    pdata = pub_bad_OTSType;
-    pdatalen = sizeof(pub_bad_OTSType);
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
+        if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
+            goto end;
+        p++;
+    }
     ret = 1;
 end:
-    EVP_PKEY_free(pkey);
     OSSL_DECODER_CTX_free(dctx);
     return ret;
+}
+
+static int lms_pubkey_decoder_fail_test(void)
+{
+    OSSL_PARAM params[6], *p = params;
+
+    *p++ = OSSL_PARAM_construct_octet_string("Invalid Small Pub Len",
+        (void *)testdata->pub, 3);
+    *p++ = OSSL_PARAM_construct_octet_string("Invalid Pub Len",
+        (void *)testdata->pub, SIZE_MAX);
+    if (testdata == lms_testdata) {
+        static const unsigned char lms_pub_bad_LMSType[] = {
+            0x00, 0x00, 0x00, 0xAA
+        };
+        static const unsigned char lms_pub_bad_OTSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+        *p++ = OSSL_PARAM_construct_octet_string("Bad LMS type",
+            (void *)lms_pub_bad_LMSType,
+            sizeof(lms_pub_bad_LMSType));
+        *p++ = OSSL_PARAM_construct_octet_string("Bad OTS type",
+            (void *)lms_pub_bad_OTSType,
+            sizeof(lms_pub_bad_OTSType));
+    } else {
+        static const unsigned char hss_pub_bad_L[] = {
+            0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x01
+        };
+        static const unsigned char hss_pub_bad_LMSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+        static const unsigned char hss_pub_bad_OTSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+
+        *p++ = OSSL_PARAM_construct_octet_string("bad L",
+            (void *)hss_pub_bad_L,
+            sizeof(hss_pub_bad_L));
+        *p++ = OSSL_PARAM_construct_octet_string("bad LMS type",
+            (void *)hss_pub_bad_LMSType,
+            sizeof(hss_pub_bad_LMSType));
+        *p++ = OSSL_PARAM_construct_octet_string("bad OTS type",
+            (void *)hss_pub_bad_OTSType,
+            sizeof(hss_pub_bad_OTSType));
+    }
+    *p = OSSL_PARAM_construct_end();
+
+    return do_pubkey_decoder_fail(params);
 }
 
 static EVP_PKEY *key_decode_from_bio(BIO *bio, const char *keytype)
@@ -132,7 +173,7 @@ static EVP_PKEY *key_decode_from_bio(BIO *bio, const char *keytype)
     int selection = 0;
 
     if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pkey, NULL, NULL,
-                      keytype,
+                      "LMS",
                       selection,
                       libctx, NULL)))
         return NULL;
@@ -161,7 +202,7 @@ static EVP_PKEY *key_decode_from_data(const unsigned char *data, size_t datalen,
 static int lms_key_decode_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td1 = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td1 = &testdata[0];
     EVP_PKEY *key = NULL;
 
     ret = TEST_ptr(key = key_decode_from_data(td1->pub, td1->publen, NULL));
@@ -198,8 +239,8 @@ static int lms_key_eq_test(void)
 {
     int ret = 0;
     EVP_PKEY *key[4];
-    LMS_ACVP_TEST_DATA *td1 = &lms_testdata[0];
-    LMS_ACVP_TEST_DATA *td2 = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td1 = &testdata[0];
+    LMS_ACVP_TEST_DATA *td2 = &testdata[1];
     size_t i;
 #ifndef OPENSSL_NO_EC
     EVP_PKEY *eckey = NULL;
@@ -208,9 +249,9 @@ static int lms_key_eq_test(void)
     for (i = 0; i < OSSL_NELEM(key); i++)
         key[i] = NULL;
 
-    if (!TEST_ptr(key[0] = lms_pubkey_from_data(td1->pub, td1->publen))
-        || !TEST_ptr(key[1] = lms_pubkey_from_data(td1->pub, td1->publen))
-        || !TEST_ptr(key[2] = lms_pubkey_from_data(td2->pub, td2->publen))
+    if (!TEST_ptr(key[0] = pubkey_from_data(td1->pub, td1->publen))
+        || !TEST_ptr(key[1] = pubkey_from_data(td1->pub, td1->publen))
+        || !TEST_ptr(key[2] = pubkey_from_data(td2->pub, td2->publen))
         || !TEST_ptr(key[3] = key_decode_from_data(td1->pub, td1->publen,
                          NULL)))
         goto end;
@@ -242,7 +283,7 @@ static int lms_key_validate_test(void)
     EVP_PKEY_CTX *vctx = NULL;
     EVP_PKEY *key = NULL;
 
-    if (!TEST_ptr(key = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(key = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(libctx, key, NULL)))
         goto end;
@@ -256,13 +297,13 @@ end:
 static int lms_verify_test(int tst)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[tst];
+    LMS_ACVP_TEST_DATA *td = &testdata[tst];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
 
-    ret = TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    ret = TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         && TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL))
         && TEST_int_eq(EVP_PKEY_verify_message_init(ctx, sig, NULL), 1)
         && TEST_int_eq(EVP_PKEY_verify(ctx, td->sig, td->siglen,
@@ -278,12 +319,12 @@ static int lms_verify_test(int tst)
 static int lms_digest_verify_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY *pub = NULL;
     EVP_MD_CTX *vctx = NULL;
     int expected = 1;
 
-    if (!TEST_ptr(pub = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(pub = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_MD_CTX_new()))
         goto err;
@@ -306,11 +347,11 @@ err:
 static int lms_digest_signing_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY *pub = NULL;
     EVP_MD_CTX *vctx = NULL;
 
-    if (!TEST_ptr(pub = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(pub = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_MD_CTX_new()))
         goto err;
@@ -328,13 +369,13 @@ err:
 static int lms_message_signing_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
 
-    ret = TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    ret = TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         && TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL))
         && TEST_int_eq(EVP_PKEY_sign_message_init(ctx, sig, NULL), -2);
 
@@ -349,7 +390,7 @@ static int lms_paramgen_fail_test(void)
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
 
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", NULL))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, NULL))
         && TEST_int_eq(EVP_PKEY_paramgen_init(ctx), -2);
 
     EVP_PKEY_CTX_free(ctx);
@@ -361,7 +402,7 @@ static int lms_keygen_fail_test(void)
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
 
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", NULL))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, NULL))
         && TEST_int_eq(EVP_PKEY_keygen_init(ctx), -2);
 
     EVP_PKEY_CTX_free(ctx);
@@ -371,11 +412,11 @@ static int lms_keygen_fail_test(void)
 static int lms_verify_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
         || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
     /* Only one shot mode is supported, streaming fails to initialise */
@@ -391,7 +432,7 @@ end:
 static int lms_verify_bad_sig_test(void)
 {
     int ret = 0, i = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
     EVP_PKEY_CTX *ctx = NULL;
@@ -412,8 +453,8 @@ static int lms_verify_bad_sig_test(void)
     if (sig_data == NULL)
         return 0;
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
 
@@ -454,7 +495,7 @@ end:
 static int lms_verify_bad_sig_len_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
     EVP_PKEY_CTX *ctx = NULL;
@@ -462,8 +503,8 @@ static int lms_verify_bad_sig_len_test(void)
     const int step = 3;
     unsigned char sigdata[4096];
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
 
@@ -500,7 +541,7 @@ end:
 
 static int lms_verify_bad_pub_sig_test(void)
 {
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     int ret = 0, i = 0;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
@@ -512,13 +553,13 @@ static int lms_verify_bad_pub_sig_test(void)
     if (!TEST_ptr(pub = OPENSSL_memdup(td->pub, td->publen)))
         return 0;
 
-    if (!TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL)))
+    if (!TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL)))
         goto end;
 
     for (i = 0; i < (int)td->publen; i += step) {
         pub[i] ^= 1; /* corrupt a byte */
         /* Corrupting the public key may cause the key load to fail */
-        pkey = lms_pubkey_from_data(pub, td->publen);
+        pkey = pubkey_from_data(pub, td->publen);
         if (pkey != NULL) {
             if (!TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
                 goto end;
@@ -556,6 +597,7 @@ const OPTIONS *test_get_options(void)
         OPT_TEST_OPTIONS_DEFAULT_USAGE,
         { "config", OPT_CONFIG_FILE, '<',
             "The configuration file to use for the libctx" },
+        { "keytype", OPT_KEY_ALG, 's', "set to either LMS or HSS" },
         { NULL }
     };
     return options;
@@ -578,6 +620,9 @@ int setup_tests(void)
             break;
         case OPT_TEST_CASES:
             break;
+        case OPT_KEY_ALG:
+            key_alg = opt_arg();
+            break;
         default:
         case OPT_ERR:
             return 0;
@@ -585,10 +630,14 @@ int setup_tests(void)
     }
     if (!test_get_libctx(&libctx, &nullprov, config_file, &libprov, NULL))
         return 0;
+    if (strcmp(key_alg, "HSS") == 0) {
+        testdata = hss_testdata;
+        testdata_len = OSSL_NELEM(hss_testdata);
+    }
 
-    ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", propq);
+    ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, propq);
     if (ctx == NULL && ERR_get_error() == EVP_R_UNSUPPORTED_ALGORITHM)
-        return TEST_skip("LMS algorithm is not available in provider");
+        return TEST_skip("HSS/LMS algorithm is not available in provider");
     EVP_PKEY_CTX_free(ctx);
 
     ADD_TEST(lms_bad_pub_len_test);
@@ -597,7 +646,7 @@ int setup_tests(void)
     ADD_TEST(lms_key_decode_test);
     ADD_TEST(lms_pubkey_decoder_test);
     ADD_TEST(lms_pubkey_decoder_fail_test);
-    ADD_ALL_TESTS(lms_verify_test, OSSL_NELEM(lms_testdata));
+    ADD_ALL_TESTS(lms_verify_test, (int)testdata_len);
     ADD_TEST(lms_verify_fail_test);
     ADD_TEST(lms_digest_verify_fail_test);
     ADD_TEST(lms_digest_signing_fail_test);